Include package dependencies in README.md
First of all - thank you for writing this library! It's exactly what i needed!

Just wanted to make a small improvement, as it took me a while to work out which package I needed to install to be able to compile and run the code.

Otherwise, it all worked great! Thanks so much!